Fixed error when selecting the three hourly forecast option
The app crashed when veiwing the three hourly forecast.
This was becasue the Forecast objects in the Three_Hourly_Forecast
object were not recieving the units key value pair in the weather dict in  get_3_hourly_forecast().
The solution was to pass the units as an arguement to get_3_hourly_forecast() and include
the units in the weather dict being passed to each Forecast object.