Bug fixed. Can select first location from list of recent locations
There was a bug preventing the user from choosing
the first location from the list of recent locations.
This was caused by the conditional on line 94
in main.py. Since the first index in the list is 0,
the condition results to false. Fixed this by checking
if option is None instead of a falsy value.